requirements.txt
slackbot==0.5.4
requests
python-dotenv